requirements.txt
tensorflow
streamlit
pillow
numpy
pandas
scikit-learn